Implemented 'open' and 'close' commands
For this sake function toggle was divided for open, close and toggle self.

Seems calling 'open' on opened object (and 'close' on closed) cause
unnecessary code to be executed, but it doesn't break anything. May
be it will be better to have asserts like 'if(collapsed) return false;'
at the begining of 'open' and 'close' functions